Update contact.py
In the view function, added error message if user inputs an ID number that isn't assigned to an existing contact. (Note: I tried to get this to create an error if the user input a number that wasn't on the list, and it kept creating problems, unfortunately, and nothing else I tried was working. So, entering 0 won't bring the error up.)

Phone number formatting can't be tailored to specific countries without either installing the phonenumber library or asking the user for their country or country code and running that against a list to format.